move deposit related logic in single file
