--- x/slashrefund/keeper/deposit_pool_test.go
@@ -1,52 +1,0 @@
-package keeper_test
-
-import (
-	"testing"
-
-	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/made-in-block/slash-refund/x/slashrefund/keeper"
-	"github.com/made-in-block/slash-refund/x/slashrefund/testslashrefund"
-	"github.com/made-in-block/slash-refund/x/slashrefund/types"
-	"github.com/stretchr/testify/require"
-)
-
-func createNDepositPool(keeper *keeper.Keeper, ctx sdk.Context, n int) []types.DepositPool {
-	items := make([]types.DepositPool, n)
-	for i := range items {
-		valPubk := secp256k1.GenPrivKey().PubKey()
-		valAddr := sdk.ValAddress(valPubk.Address())
-		items[i].OperatorAddress = valAddr.String()
-		items[i].Shares = sdk.NewDec(int64(1000 * i))
-		items[i].Tokens = sdk.NewInt64Coin("stake", int64(1000*i))
-		keeper.SetDepositPool(ctx, items[i])
-	}
-	return items
-}
-
-func TestDepositPoolGet(t *testing.T) {
-	keeper, ctx := testslashrefund.NewTestKeeper(t)
-	items := createNDepositPool(keeper, ctx, 10)
-	for _, item := range items {
-		valAddr, _ := sdk.ValAddressFromBech32(item.OperatorAddress)
-		rst, found := keeper.GetDepositPool(ctx, valAddr)
-		require.True(t, found)
-		require.Equal(t, item, rst)
-	}
-}
-func TestDepositPoolRemove(t *testing.T) {
-	keeper, ctx := testslashrefund.NewTestKeeper(t)
-	items := createNDepositPool(keeper, ctx, 10)
-	for _, item := range items {
-		valAddr, _ := sdk.ValAddressFromBech32(item.OperatorAddress)
-		keeper.RemoveDepositPool(ctx, valAddr)
-		_, found := keeper.GetDepositPool(ctx, valAddr)
-		require.False(t, found)
-	}
-}
-
-func TestDepositPoolGetAll(t *testing.T) {
-	keeper, ctx := testslashrefund.NewTestKeeper(t)
-	items := createNDepositPool(keeper, ctx, 10)
-	require.ElementsMatch(t, items, keeper.GetAllDepositPool(ctx))
-}
